% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_issue.R
\name{draft_issue}
\alias{draft_issue}
\title{Create a GitHub issue from Rmarkdown}
\usage{
draft_issue(
  filename = "issue.Rmd",
  tempdir = getOption("draft_issue_in_tempdir", TRUE),
  overwrite = TRUE
)
}
\arguments{
\item{filename}{the Rmarkdown file name to contain your issue}

\item{tempdir}{use a temporary directory if TRUE, current working directory if FALSE}

\item{overwrite}{whether or not to overwrite an existing issue with the same filename (defaults to TRUE).}
}
\description{
An Rmarkdown document is created an opened that can add or comment on
issues/PRs when rendered with \code{rmarkdown::render()}, the 'knit' button in
RStudio, or the 'Knit Rmd' command in VSCode.

By default issues are created in a temporary dir, but this can be overidden
and they will be created in the current directory.
}
