% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_issue.R
\name{github_issue}
\alias{github_issue}
\title{A github issue Rmarkdown format}
\usage{
github_issue(
  repo = NULL,
  number = NULL,
  labels = NULL,
  action = "create",
  draft = TRUE,
  close_with_comment = FALSE,
  fig_width = 7,
  fig_height = 5,
  dev = "png",
  df_print = "default",
  math_method = "default",
  wrap = "preserve"
)
}
\arguments{
\item{repo}{the repository create the issue on e.g. "milesmcbain/capsule", or
"capsule". Name will be resolved against locally insalled packages and CRAN in
the later case.}

\item{number}{the issue number in the repository if performing a comment or update action}

\item{labels}{the labels to set for the issue if performing create or update action}

\item{action}{the type of action to perform: "create", "update", "comment". Only "comment" is valid for PRs.}

\item{draft}{if TRUE the action is not performed after rendering the document to
GitHub markdown - set to TRUE initially to give you a chance to preview
markdown output.}

\item{close_with_comment}{if TRUE close the issue thread with comment action (assuming you have appropriate repo permissions).}

\item{fig_width}{passed to image dimension passed to \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}}}

\item{fig_height}{image dimension passed to image dimension passed to \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}}}

\item{dev}{graphics device argument passed to \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}}}

\item{df_print}{data.frame print method passed to \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}}}

\item{math_method}{latex style math expression rendering method passed \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}}}

\item{wrap}{argument passed to pandoc. Controls use of line breaks. One of
"auto", "preserve" or "none". "preserve" will mostly not interfere with your
text formatting.}
}
\description{
A wraper for \code{\link[rmarkdown:github_document]{rmarkdown::github_document()}} that can peform actions on GitHub
when rendering relating to issues and PRs including:
\itemize{
\item create an issue
\item update an issue title, body, and labels
\item comment on an issue
\item close an issue with a comment
}
}
