% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuecreep-package.R
\docType{package}
\name{issuecreep-package}
\alias{issuecreep}
\alias{issuecreep-package}
\title{issuecreep: R to GitHub Rmarkdown productivity}
\description{
Browse and interact with GitHub via an Rmarkdown document interface.
}
\author{
\strong{Maintainer}: Miles McBain \email{miles.mcbain@gmail.com} (\href{https://orcid.org/0000-0003-2865-2548}{ORCID})

}
\keyword{internal}
