% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_issue.R
\name{save_issue}
\alias{save_issue}
\title{Save an issue to the current working directory}
\usage{
save_issue(
  filename = NULL,
  folder = getOption("rmdgh_issue_location", "./issues")
)
}
\arguments{
\item{filename}{the name to give the issue. Defaulted from issue repo and number in yaml.}

\item{folder}{where to store the issue. The default is "./issues".}
}
\description{
A helper for collecting issue documents locally to worked on at a later time.
}
\details{
\code{folder} will be \code{./issues} by default but is configurable in option
\code{rmdgh_issue_location}. If it does not exist, it is created and added to the
\code{.Rbuildignore}.
}
